requirements_uwsgi.txt
alembic==1.7.7
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
beautifulsoup4==4.10.0
bidict==0.22.0
blinker==1.4
boto3==1.24.15
botocore==1.27.15
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.4
cryptography==37.0.4
Deprecated==1.2.13
dnspython==2.2.1
elastic-transport==8.1.0
elasticsearch==8.1.1
eventlet==0.33.1
fastjsonschema==2.15.3
Flask==2.0.3
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Injector==0.13.0
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Seeder==1.2.0
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
injector==0.19.0
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.1
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.1
redis==4.1.4
requests==2.28.1
s3transfer==0.6.0
sentry-sdk==1.5.8
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
urllib3==1.26.9
uWSGI==2.0.20
watchdog==2.1.7
Werkzeug==2.0.3
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0